CSIPOAssetCreationTool/DemoVideo/requirements.txt
moviepy>=2.0
Pillow>=10.0
elevenlabs>=1.0
python-dotenv>=1.0
numpy>=1.24
